Move ModuleHybridEngine to its own file
